apply triggertemplate param default value if body/header based triggerbinding param missing from body/header
